[refactor] #171 - MyPageService의 친구추천 로직 중 공통되는 부분 RandomUserPicker, UserCalculator 객체로 분리
